+++ src/com/koxudaxi/pydantic/PydanticDataclassTypeProvider.kt
@@ -4,7 +4,6 @@ import com.intellij.psi.PsiElement
 import com.jetbrains.python.codeInsight.stdlib.PyDataclassTypeProvider
 import com.jetbrains.python.psi.*
 import com.jetbrains.python.psi.impl.PyCallExpressionImpl
-import com.jetbrains.python.psi.impl.PyCallExpressionNavigator
 import com.jetbrains.python.psi.types.*
 
 /**
@@ -18,84 +17,11 @@ import com.jetbrains.python.psi.types.*
  *
  */
 class PydanticDataclassTypeProvider : PyTypeProviderBase() {
-    private val pyDataclassTypeProvider = PyDataclassTypeProvider()
-    private val pydanticTypeProvider = PydanticTypeProvider()
-    override fun getReferenceExpressionType(
-        referenceExpression: PyReferenceExpression,
-        context: TypeEvalContext,
-    ): PyType? {
-        return getPydanticDataclass(
-            referenceExpression,
-            TypeEvalContext.codeInsightFallback(referenceExpression.project)
-        )
-    }
-
-
-    internal fun getDataclassCallableType(
-        referenceTarget: PsiElement,
-        context: TypeEvalContext,
-        callSite: PyCallExpression?,
-    ): PyCallableType? {
-        return pyDataclassTypeProvider.getReferenceType(
-            referenceTarget,
-            context,
-            callSite ?: PyCallExpressionImpl(referenceTarget.node)
-        )?.get() as? PyCallableType
-    }
-
-    private fun getPydanticDataclassType(
-        referenceTarget: PsiElement,
-        context: TypeEvalContext,
-        pyReferenceExpression: PyReferenceExpression,
-        definition: Boolean,
-    ): PyType? {
-        val callSite = PyCallExpressionNavigator.getPyCallExpressionByCallee(pyReferenceExpression)
-        val dataclassCallableType = getDataclassCallableType(referenceTarget, context, callSite) ?: return null
-
-        val dataclassType = (dataclassCallableType).getReturnType(context) as? PyClassType ?: return null
-        if (!dataclassType.pyClass.isPydanticDataclass) return null
-        val ellipsis = PyElementGenerator.getInstance(referenceTarget.project).createEllipsis()
-        val injectedPyCallableType = PyCallableTypeImpl(
-            dataclassCallableType.getParameters(context)?.map {
-                when {
-                    it.defaultValueText == "..." && it.defaultValue is PyNoneLiteralExpression ->
-                        pydanticTypeProvider.injectDefaultValue(dataclassType.pyClass, it, ellipsis, null, context)
-                            ?: it
-
-                    else -> it
-                }
-            }, dataclassType
-        )
-        val injectedDataclassType = (injectedPyCallableType).getReturnType(context) as? PyClassType ?: return null
-        return when {
-            callSite is PyCallExpression && definition -> injectedPyCallableType
-            definition -> injectedDataclassType.toClass()
-            else -> injectedDataclassType
+    override fun getCallableType(callable: PyCallable, context: TypeEvalContext): PyType? {
+        if (callable is PyFunction && callable.isPydanticDataclass) {
+            // Drop fake dataclass return type
+            return PyCallableTypeImpl(callable.getParameters(context), null)
         }
-    }
-
-    private fun getPydanticDataclass(referenceExpression: PyReferenceExpression, context: TypeEvalContext): PyType? {
-        return getResolvedPsiElements(referenceExpression, context)
-            .asSequence()
-            .mapNotNull {
-                when {
-                    it is PyClass && it.isPydanticDataclass ->
-                        getPydanticDataclassType(it, context, referenceExpression, true)
-
-                    it is PyTargetExpression -> (it as? PyTypedElement)
-                        ?.getType(context)?.pyClassTypes
-                        ?.filter { pyClassType -> pyClassType.pyClass.isPydanticDataclass }
-                        ?.firstNotNullOfOrNull { pyClassType ->
-                            getPydanticDataclassType(
-                                pyClassType.pyClass,
-                                context,
-                                referenceExpression,
-                                pyClassType.isDefinition
-                            )
-                        }
-
-                    else -> null
-                }
-            }.firstOrNull()
+        return super.getCallableType(callable, context)
     }
 }